requirements.txt
asgiref==3.7.2
Django==4.2.6
django-leaflet==0.29.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
pytz==2023.3.post1
sqlparse==0.4.4
django-environ==0.11.2
psycopg2-binary==2.9.9
django-filter==2.4.0